+++ NEWS.md
@@ -16,4 +16,8 @@
     * website update, added coc and contribution [2022-12-13]
     * website update, coc links (and minor fixes) [2022-12-14]
     * website update, news update [2022-12-14]
+    * generalised ef_beddows... functions and extended to pm10 [2022-12-14] 
+* [0.0.4]
+    * added vehicle/fleet object workforce function [2022-12-21]
     
+